move panoptic tests to nightly and add to ci
