+++ .github/workflows/tt-metal-l2-nightly.yaml
@@ -40,6 +40,11 @@ on:
         required: false
         type: boolean
         default: false
+      run_bos_tests:
+        description: 'Run TTNN BOS tests'
+        required: false
+        type: boolean
+        default: false
       timeout:
         description: 'Test timeout in minutes'
         required: false
@@ -84,6 +89,7 @@ jobs:
       run_sdxl_tests: ${{ github.event_name == 'schedule' || inputs.run_sdxl_tests }}
       run_train_tests: ${{ github.event_name == 'schedule' || inputs.run_train_tests }}
       run_sdpa_stress_tests: ${{ github.event_name == 'schedule' || inputs.run_sdpa_stress_tests }}
+      run_ttnn-bos-tests: ${{ github.event_name == 'schedule' || inputs.run_bos_tests }}
   didt-tests:
     if: ${{ github.event_name == 'schedule' || inputs.run_didt_tests }}
     needs: build-artifact
