+++ tests/ttnn/nightly/unit_tests/operations/conv/test_conv2d.py
@@ -11,7 +11,8 @@ from models.utility_functions import (
     is_wormhole_b0,
     is_blackhole,
 )
-from models.utility_functions import skip_for_blackhole
+from models.utility_functions import skip_for_blackhole, run_for_blackhole
+from tests.ttnn.unit_tests.test_bh_20_cores_sharding import skip_if_not_blackhole_20_cores
 from tests.ttnn.utils_for_testing import assert_with_pcc, check_with_pcc_without_tensor_printout
 import ttnn
 from ttnn.operations.conv2d import get_torch_act_func_from_string
@@ -19,6 +20,13 @@ from ttnn.operations.conv2d import get_torch_act_func_from_string
 HS = ttnn.TensorMemoryLayout.HEIGHT_SHARDED
 BS = ttnn.TensorMemoryLayout.BLOCK_SHARDED
 WS = ttnn.TensorMemoryLayout.WIDTH_SHARDED
+
+try:
+    from tracy import signpost
+except ImportError:
+    # Fallback implementation if tracy module is not available
+    def signpost(*args, **kwargs):
+        pass
 
 
 def torch_fast_pcc(golden, calculated, pcc=0.99):
@@ -356,7 +364,7 @@ def run_conv_with_split(
     stride_h,
     stride_w,
     padding,
-    dilation=None,
+    dilation=(1, 1),
     config_override={},
     shard_layout=None,
     split_input_channels_factor=2,
@@ -1440,6 +1448,7 @@ def test_sd_conv(
             stride_h,
             stride_w,
             (pad_h, pad_w),
+            (1, 1),
             config_override,
             shard_layout=shard_layout,
             split_input_channels_factor=3 if input_channels == 1920 else 2,
@@ -1592,6 +1601,7 @@ def test_sd_conv_wh(
             stride_h,
             stride_w,
             (pad_h, pad_w),
+            (1, 1),
             config_override,
             shard_layout=shard_layout,
             split_input_channels_factor=3 if input_channels == 1920 else 2,
@@ -1675,7 +1685,8 @@ def test_sd14_vae_conv(
             stride[0],
             stride[1],
             padding,
-            None,
+            (1, 1),
+            {},
             shard_layout=None,
             split_input_channels_factor=split_factor_input_channels,
             split_output_channels_factor=split_factor_output_channels,
@@ -3229,6 +3240,7 @@ def test_conv2d_sdxl(
             stride[0],
             stride[1],
             padding,
+            dilation,
             config_override,
             shard_layout=shard_layout,
             split_input_channels_factor=split_input_channels_factor,
@@ -3380,6 +3392,7 @@ def test_conv2d_sdxl_refiner(
             stride[0],
             stride[1],
             padding,
+            dilation,
             config_override,
             shard_layout=shard_layout,
             split_input_channels_factor=split_input_channels_factor,
@@ -4209,3 +4222,267 @@ def test_conv_sharded_rm_input(
         enable_act_double_buffer=act_db,
         enable_weights_double_buffer=w_db
     )
+
+@pytest.mark.parametrize(
+    "batch_size, input_channels, output_channels, input_height, input_width, filter_height, filter_width, stride_h, stride_w, pad_h, pad_w, groups, has_bias, multipler",
+    [
+        (1, 1024, 2048, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 1),
+        (1, 1024, 256, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 5),
+        (1, 1024, 512, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 1),
+        (1, 128, 128, 128, 256, 3, 3, 1, 1, 1, 1, 1, False, 3),
+        (1, 128, 128, 128, 256, 3, 3, 2, 2, 1, 1, 1, False, 1),
+        (1, 128, 256, 128, 256, 1, 1, 1, 1, 0, 0, 1, False, 1),
+        (1, 128, 32, 128, 256, 3, 3, 1, 1, 1, 1, 1, False, 2),
+        (1, 128, 64, 128, 256, 1, 1, 1, 1, 0, 0, 1, False, 1),
+        (1, 128, 128, 64, 128, 3, 3, 1, 1, 1, 1, 1, False, 4),
+        (1, 128, 512, 64, 128, 1, 1, 1, 1, 0, 0, 1, False, 4),
+        (1, 1280, 256, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 2),
+        (1, 160, 128, 128, 256, 3, 3, 1, 1, 1, 1, 1, False, 1),
+        (1, 2048, 256, 1, 1, 1, 1, 1, 1, 0, 0, 1, True, 2),
+        (1, 2048, 256, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 2),
+        (1, 2048, 512, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 2),
+        (1, 256, 128, 128, 256, 1, 1, 1, 1, 0, 0, 1, False, 1),
+        (1, 256, 19, 128, 256, 1, 1, 1, 1, 0, 0, 1, True, 1),
+        (1, 256, 32, 128, 256, 1, 1, 1, 1, 0, 0, 1, False, 2),
+        (1, 256, 64, 128, 256, 1, 1, 1, 1, 0, 0, 1, False, 2),
+        (1, 256, 1024, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 6),
+        (1, 256, 256, 32, 64, 3, 3, 1, 1, 1, 1, 1, False, 5),
+        (1, 256, 256, 64, 128, 3, 3, 2, 2, 1, 1, 1, False, 1),
+        (1, 256, 256, 64, 128, 3, 3, 1, 1, 1, 1, 1, False, 1),
+        (1, 3, 64, 512, 1024, 3, 3, 2, 2, 1, 1, 1, False, 1),
+        (1, 32, 1, 128, 256, 1, 1, 1, 1, 0, 0, 1, True, 1),
+        (1, 32, 2, 128, 256, 1, 1, 1, 1, 0, 0, 1, True, 1),
+        (1, 320, 128, 64, 128, 3, 3, 1, 1, 1, 1, 1, False, 1),
+        (1, 320, 256, 64, 128, 3, 3, 1, 1, 1, 1, 1, False, 1),
+        (1, 512, 2048, 32, 64, 1, 1, 1, 1, 0, 0, 1, False, 3),
+        (1, 512, 512, 32, 64, 3, 3, 1, 1, 2, 2, 1, False, 1),
+        (1, 512, 512, 32, 64, 3, 3, 1, 1, 4, 4, 1, False, 1),
+        (1, 512, 512, 32, 64, 3, 3, 1, 1, 8, 8, 1, False, 1),
+        (1, 512, 1024, 64, 128, 1, 1, 2, 2, 0, 0, 1, False, 1),
+        (1, 512, 128, 64, 128, 1, 1, 1, 1, 0, 0, 1, False, 3),
+        (1, 512, 256, 64, 128, 1, 1, 1, 1, 0, 0, 1, False, 1),
+        (1, 512, 64, 64, 128, 1, 1, 1, 1, 0, 0, 1, False, 2),
+        (1, 64, 256, 128, 256, 1, 1, 1, 1, 0, 0, 1, False, 3),
+        (1, 64, 64, 128, 256, 3, 3, 1, 1, 1, 1, 1, False, 3),
+    ],
+)
+@pytest.mark.parametrize("device_params", [{"l1_small_size": 32768}], indirect=True)
+@run_for_blackhole("blackhole specific tests")
+def test_conv2d_panoptic(
+    device,
+    batch_size,
+    input_channels,
+    output_channels,
+    input_height,
+    input_width,
+    filter_height,
+    filter_width,
+    stride_h,
+    stride_w,
+    pad_h,
+    pad_w,
+    groups,
+    torch_tensor_map,
+    has_bias,
+    multipler,
+):
+    skip_if_not_blackhole_20_cores(device)
+
+    signpost(header=f"conv2d_{input_channels}_{output_channels}_{input_height}_{input_width}; multipler={multipler}")
+    run_conv(
+        device=device,
+        config_override=None,
+        torch_tensor_map=torch_tensor_map,
+        math_fidelity=ttnn.MathFidelity.HiFi4,
+        output_dtype=ttnn.bfloat16,
+        weights_dtype=ttnn.bfloat8_b,
+        batch_size=batch_size,
+        output_channels=output_channels,
+        input_channels=input_channels,
+        input_height=input_height,
+        input_width=input_width,
+        filter_height=filter_height,
+        filter_width=filter_width,
+        stride_h=stride_h,
+        stride_w=stride_w,
+        padding=(pad_h, pad_w),
+        groups=groups,
+        has_bias=has_bias,
+    )
+    signpost(header="conv2d_end.")
+
+
+@pytest.mark.parametrize(
+    "input_layout, dtype",
+    [
+        [ttnn.TILE_LAYOUT, ttnn.bfloat8_b],
+    ],
+)
+@pytest.mark.parametrize("device_params", [{"l1_small_size": 32768}], indirect=True)
+@pytest.mark.parametrize(
+    "batch_size, input_channels, output_channels, input_height, input_width, slice_type, num_slices, weights_dtype, kernel, stride, padding, dilation, act_block_h_override,  math_fidelity, multipler",
+    # fmt: off
+    (
+        (1, 256, 256,  128,    256,   SliceHeight,   2,  ttnn.bfloat8_b,  (3, 3), (1, 1), (1, 1), (1, 1), 32,  ttnn.MathFidelity.HiFi4, 3), # Panoptic
+        (1, 256, 512,  128,    256,   SliceHeight,   4,  ttnn.bfloat8_b,  (1, 1), (2, 2), (1, 1), (1, 1), 32,  ttnn.MathFidelity.HiFi4, 1), # Panoptic
+        (1, 288, 256,  128,    256,   SliceHeight,   4,  ttnn.bfloat8_b,  (3, 3), (1, 1), (1, 1), (1, 1), 32,  ttnn.MathFidelity.HiFi4, 1), # Panoptic
+        (1,  64, 128,  256,    512,   SliceHeight,   2,  ttnn.bfloat8_b,  (3, 3), (1, 1), (1, 1), (1, 1), 32,  ttnn.MathFidelity.HiFi4, 1), # Panoptic
+        (1,  64,  64,  256,    512,   SliceHeight,   2,  ttnn.bfloat8_b,  (3, 3), (1, 1), (1, 1), (1, 1), 32,  ttnn.MathFidelity.HiFi4, 1), # Panoptic
+    )
+    # fmt: on
+)
+@pytest.mark.parametrize(
+    "has_bias, fp32_accum, packer_l1_acc",
+    [[False, False, False]],
+)
+@run_for_blackhole("blackhole specific tests")
+def test_conv_dram_panoptic(
+    device,
+    torch_tensor_map,
+    batch_size,
+    output_channels,
+    input_channels,
+    input_height,
+    input_width,
+    has_bias,
+    weights_dtype,
+    dtype,
+    slice_type,
+    num_slices,
+    kernel,
+    stride,
+    padding,
+    dilation,
+    act_block_h_override,
+    math_fidelity,
+    fp32_accum,
+    input_layout,
+    packer_l1_acc,
+    multipler,
+):
+    skip_if_not_blackhole_20_cores(device)
+
+    config = {
+        "act_block_h": act_block_h_override,
+    }
+
+    signpost(
+        header=f"dram_slice_conv_{slice_type}_{num_slices}_slices_{input_channels}_{output_channels}_{input_height}_{input_width}; multipler={multipler}"
+    )
+    run_conv(
+        device,
+        torch_tensor_map,
+        math_fidelity,
+        dtype,
+        weights_dtype,
+        batch_size,
+        output_channels,
+        input_channels,
+        input_height,
+        input_width,
+        kernel[0],
+        kernel[1],
+        stride[0],
+        stride[1],
+        padding,
+        config,
+        deallocate_activation=True,
+        has_bias=True,
+        fp32_accum=fp32_accum,
+        packer_l1_acc=packer_l1_acc,
+        input_dtype=dtype,
+        input_layout=input_layout,
+        output_layout=input_layout,
+        run_twice=False,
+        fast_compare=False,
+        slice_config=ttnn.Conv2dSliceConfig(
+            slice_type=slice_type,
+            num_slices=num_slices,
+        ),
+    )
+    signpost(header=f"dram_slice_conv_{slice_type}_{num_slices}_slices_end.")
+
+
+@pytest.mark.parametrize(
+    "batch, input_channels, output_channels, input_height, input_width, weights_dtype, output_dtype, groups, kernel, stride, padding, dilation, shard_layout, act_block_h_override, act_block_w_div, deallocate_activation, math_fidelity, fp32_accum, packer_l1_acc, enable_split_reader, split_input_channels_factor, split_output_channels_factor, act_db, w_db, multipler",
+    # fmt: off
+    (
+        (1, 2048, 256, 32, 64, ttnn.bfloat8_b, ttnn.bfloat16, 1, (3, 3), (1, 1), (12, 12), (1, 1), None, 0, 1, False, ttnn.MathFidelity.LoFi, False, False, False, 2, 1, False, True, 2),  # Panoptic
+        (1, 2048, 256, 32, 64, ttnn.bfloat8_b, ttnn.bfloat16, 1, (3, 3), (1, 1), (18, 18), (1, 1), None, 0, 1, False, ttnn.MathFidelity.LoFi, False, False, False, 2, 1, False, True, 2),  # Panoptic
+        (1, 2048, 256, 32, 64, ttnn.bfloat8_b, ttnn.bfloat16, 1, (3, 3), (1, 1), (6, 6), (1, 1), None, 0, 1, False, ttnn.MathFidelity.LoFi, False, False, False, 2, 1, False, True, 2),  # Panoptic
+    ),
+    # fmt: on
+)
+@pytest.mark.parametrize("device_params", [{"l1_small_size": 4 * 16384}], indirect=True)
+@run_for_blackhole("blackhole specific tests")
+def test_conv2d_ch_split_dram_panoptic(
+    device,
+    torch_tensor_map,
+    batch,
+    input_channels,
+    output_channels,
+    input_height,
+    input_width,
+    weights_dtype,
+    output_dtype,
+    groups,
+    kernel,
+    stride,
+    padding,
+    dilation,
+    shard_layout,
+    act_block_h_override,
+    act_block_w_div,
+    deallocate_activation,
+    math_fidelity,
+    fp32_accum,
+    packer_l1_acc,
+    enable_split_reader,
+    split_input_channels_factor,
+    split_output_channels_factor,
+    act_db,
+    w_db,
+    multipler,
+):
+    skip_if_not_blackhole_20_cores(device)
+
+    config_override = {}
+    config_override["act_block_h"] = act_block_h_override
+    config_override["act_block_w_div"] = act_block_w_div
+
+    signpost(
+        header=f"ch_slice_conv_{split_input_channels_factor}_{split_output_channels_factor}_x_{input_channels}_{output_channels}_{input_height}_{input_width}; multipler={multipler}"
+    )
+
+    if split_input_channels_factor > 1 or split_output_channels_factor > 1:
+        run_conv_with_split(
+            device,
+            torch_tensor_map,
+            math_fidelity,
+            output_dtype,
+            weights_dtype,
+            batch,
+            output_channels,
+            input_channels,
+            input_height,
+            input_width,
+            kernel[0],
+            kernel[1],
+            stride[0],
+            stride[1],
+            padding,
+            dilation,
+            config_override,
+            shard_layout=shard_layout,
+            split_input_channels_factor=split_input_channels_factor,
+            split_output_channels_factor=split_output_channels_factor,
+            fp32_accum=fp32_accum,
+            packer_l1_acc=packer_l1_acc,
+            auto_shard=True if shard_layout is None else False,
+            input_layout=ttnn.TILE_LAYOUT if output_dtype == ttnn.bfloat8_b else None,
+            enable_act_double_buffer=act_db,
+            enable_weights_double_buffer=w_db,
+        )
+    else:
+        pytest.skip("Not a split conv test, skipping.")
+    signpost(header=f"ch_slice_conv_{split_input_channels_factor}_{split_output_channels_factor}_end.")
